- allow empty group_info_table in re-run mode - add error hint when node identity is different from the database
